MAGE-1095: create index builders and move logic from DataHelper
